Add a link to the current survey on the landing page
When a visitor goes to the landing page now until the survey is over, they’ll get a link to the current survey and see the YouTube video in which we describe each item. This is so they can immediately fill out the survey and sign up for our email list without having to wait for me to process their entry through the normal form.

After the current survey is over, the page will go back to how it looked before, with our photo added to the end of the page.

I realize there’s a lot of room for improvement with the design. I think this is sufficient until we get additional design help, although of course I am open to any feedback about how we can improve the landing page.