Add styles to advertise the community survey
Similar to how the survey ad works on canjs.com, the control shows up at the bottom of the viewport after the user has visited three pages (ever). There’s an X button to close it, but otherwise the entire bar is a link to the survey landing page. The bar shows up on every single page except the survey page itself.

The bar shows on top of most content but under the left navigation. It does not show up on mobile.

Related to https://github.com/donejs/donejs/issues/1019